+++ packages/core/src/common/debug.c
@@ -577,7 +577,7 @@ static void cheatAllItems(PlayState* play)
     gMmExtraItems.ocarina = 3;
     gMmExtraItems.hammerGFS = 3;
 
-    gCustomSave.halfDays = 0b111111;
+    // gCustomSave.halfDays = 0b111111;
 
     gSave.info.inventory.items[ITS_MM_TRADE1] = ITEM_MM_DEED_LAND;
     gSave.info.inventory.items[ITS_MM_TRADE2] = ITEM_MM_ROOM_KEY;